deeplens/requirements.txt
absl-py==0.5.0
astor==0.7.1
astroid==1.6.5
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
configparser==3.5.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
isort==4.3.4
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
lazy-object-proxy==1.3.1
Markdown==3.0
mccabe==0.6.1
mock==2.0.0
numpy==1.14.5
pbr==4.2.0
Pillow==5.2.0
protobuf==3.6.1
pylint==1.9.3
PyYAML==3.13
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
